+++ nequip/datasets/aspirin.py
@@ -2,14 +2,11 @@ import numpy as np
 
 from os.path import dirname, basename, abspath
 
-from ase import units
-from ase.io import read
-
 from nequip.data import AtomicDataDict, AtomicInMemoryDataset
 
 
 class AspirinDataset(AtomicInMemoryDataset):
-    """Aspirin DFT/CCSD(T) data """
+    """Aspirin DFT/CCSD(T) data"""
 
     URL = "http://quantum-machine.org/gdml/data/npz/aspirin_ccsd.zip"
     FILE_NAME = "benchmark_data/aspirin_ccsd-train.npz"
@@ -30,8 +27,7 @@ class AspirinDataset(AtomicInMemoryDataset):
             AtomicDataDict.TOTAL_ENERGY_KEY: data["E"].reshape([-1, 1]),
         }
         fixed_fields = {
-            AtomicDataDict.ATOMIC_NUMBERS_KEY: np.asarray(data["z"], dtype=np.int),
+            AtomicDataDict.ATOMIC_NUMBERS_KEY: np.asarray(data["z"], dtype=int),
             AtomicDataDict.PBC_KEY: np.array([False, False, False]),
-            AtomicDataDict.CELL_KEY: None,
         }
         return arrays, fixed_fields