+++ nequip/scripts/train.py
@@ -28,6 +28,7 @@ default_config = dict(
     model_initializers=[],
     dataset_statistics_stride=1,
     default_dtype="float32",
+    allow_tf32=True,
     verbose="INFO",
     model_debug_mode=False,
     equivariance_test=False,
@@ -79,6 +80,14 @@ def _load_callable(obj: Union[str, Callable]) -> Callable:
 
 def fresh_start(config):
     # = Set global state =
+    # Set TF32 support
+    # See https://pytorch.org/docs/stable/notes/cuda.html#tensorfloat-32-tf32-on-ampere-devices
+    if torch.cuda.is_available():
+        if torch.torch.backends.cuda.matmul.allow_tf32 and not config.allow_tf32:
+            # it is enabled, and we dont want it to, so disable:
+            torch.backends.cuda.matmul.allow_tf32 = False
+            torch.backends.cudnn.allow_tf32 = False
+
     if config.model_debug_mode:
         set_irreps_debug(enabled=True)
     torch.set_default_dtype(
@@ -216,6 +225,11 @@ def fresh_start(config):
         + (
             [AtomicDataDict.FORCE_KEY]
             if AtomicDataDict.FORCE_KEY in core_model.irreps_out
+            else []
+        )
+        + (
+            [AtomicDataDict.PER_ATOM_ENERGY_KEY]
+            if AtomicDataDict.PER_ATOM_ENERGY_KEY in core_model.irreps_out
             else []
         ),
         scale_by=global_scale,