+++ tests/nn/test_sequential.py
@@ -38,3 +38,30 @@ def test_append():
         }
     )
     assert out["thing"].shape == out[AtomicDataDict.NODE_FEATURES_KEY].shape
+
+
+def test_insert():
+    sgn = SequentialGraphNetwork.from_parameters(
+        shared_params={"num_species": 3},
+        layers={"one_hot": OneHotAtomEncoding, "lin2": AtomwiseLinear},
+    )
+    sgn.insert_from_parameters(
+        after="one_hot",
+        shared_params={"out_field": "thing"},
+        name="lin1",
+        builder=AtomwiseLinear,
+        params={"out_field": AtomicDataDict.NODE_FEATURES_KEY},
+    )
+    assert isinstance(sgn.lin1, AtomwiseLinear)
+    assert len(sgn) == 3
+    assert sgn[0] is sgn.one_hot
+    assert sgn[1] is sgn.lin1
+    assert sgn[2] is sgn.lin2
+    out = sgn(
+        {
+            AtomicDataDict.POSITIONS_KEY: torch.randn(5, 3),
+            AtomicDataDict.EDGE_INDEX_KEY: torch.LongTensor([[0, 1], [1, 0]]),
+            AtomicDataDict.SPECIES_INDEX_KEY: torch.LongTensor([0, 0, 1, 2, 0]),
+        }
+    )
+    assert AtomicDataDict.NODE_FEATURES_KEY in out